rpa_invoice_system/requirements.txt
# Core
python-dotenv==1.0.1
apscheduler==3.10.4
watchdog==4.0.0

# OCR & PDF Extraction
pdfplumber==0.11.0
pytesseract==0.3.10
Pillow==10.3.0
pdf2image==1.17.0

# Data Processing
pandas==2.2.2
openpyxl==3.1.2
numpy==1.26.4

# Database
SQLAlchemy==2.0.30
alembic==1.13.1

# Web Dashboard
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.3

# Report Generation
reportlab==4.2.0

# Email (IMAP)
IMAPClient==3.0.3

# NLP / Text processing
python-dateutil==2.9.0
fuzzywuzzy==0.18.0
python-Levenshtein==0.25.1

# Testing
pytest==8.2.0
pytest-cov==5.0.0
